+++ lib/src/data/data_source/local_data_source/local_storage_stub.dart
@@ -49,4 +49,3 @@ class LocalStorageImpl {
     throw UnsupportedError('LocalStorage is not supported on web/WASM');
   }
 }
-